requirements.txt
PyQt6>=6.4.0
plexapi>=4.15.0
requests
setuptools
py2app; sys_platform == 'darwin'
pyobjc-framework-MediaPlayer; sys_platform == 'darwin'
pyobjc-framework-Cocoa; sys_platform == 'darwin'
pyinstaller; sys_platform == 'win32'
pywin32; sys_platform == 'win32'
